feat: UML 기준으로 String extension 추가
target 을 기준으로 split 해준 후 substring배열을 String 배열로 바꾸어주는 extension 추가 및 사소한 오타 수정

+++ Calculator/Protocols/CalculateItem.swift
@@ -11,6 +11,6 @@ protocol CalculateItem {
     
 }
 
-extension Int: CalculateItem {
+extension Double: CalculateItem {
     
 }
